e2etest: move vector_store_cluster from validator-tests
This commit moves interface of vector_store_cluster actor from validator-test
into e2etest-vector_store_cluster. It also updates crate references.
