+++ crates/validator/src/auth.rs
@@ -8,13 +8,13 @@ use crate::common;
 use crate::common::*;
 use async_backtrace::framed;
 use e2etest_scylla_cluster::ScyllaClusterExt;
+use e2etest_vector_store_cluster::VectorStoreClusterExt;
 use std::sync::LazyLock;
 use std::time::Duration;
 use tokio::time::sleep;
 use tracing::info;
 use uuid::Uuid;
 use vector_search_validator_tests::TestCase;
-use vector_search_validator_tests::VectorStoreClusterExt;
 use vector_store::httproutes::IndexStatus;
 use vector_store::httproutes::NodeStatus;
