+++ crates/validator/src/common.rs
@@ -11,6 +11,8 @@ use e2etest_scylla_cluster::ScyllaNodeConfig;
 use e2etest_scylla_proxy_cluster::ScyllaProxyClusterExt;
 use e2etest_scylla_proxy_cluster::ScyllaProxyNodeConfig;
 use e2etest_tls::TlsExt;
+use e2etest_vector_store_cluster::VectorStoreClusterExt;
+use e2etest_vector_store_cluster::VectorStoreNodeConfig;
 use httpclient::HttpClient;
 use itertools::Itertools;
 use scylla::client::session::Session;
@@ -28,8 +30,6 @@ use std::time::Duration;
 use tap::Pipe;
 use tokio::time;
 use tracing::info;
-use vector_search_validator_tests::VectorStoreClusterExt;
-use vector_search_validator_tests::VectorStoreNodeConfig;
 use vector_store::IndexInfo;
 use vector_store::IndexName;
 use vector_store::KeyspaceName;
