+++ crates/validator/src/reconnect.rs
@@ -9,6 +9,7 @@ use async_backtrace::framed;
 use e2etest_firewall::FirewallExt;
 use e2etest_scylla_cluster::ScyllaClusterExt;
 use e2etest_scylla_proxy_cluster::ScyllaProxyClusterExt;
+use e2etest_vector_store_cluster::VectorStoreClusterExt;
 use scylla_proxy::Condition;
 use scylla_proxy::Reaction;
 use scylla_proxy::RequestReaction;
@@ -17,7 +18,6 @@ use std::time::Duration;
 use tap::Pipe;
 use tracing::info;
 use vector_search_validator_tests::TestCase;
-use vector_search_validator_tests::VectorStoreClusterExt;
 use vector_store::httproutes::IndexStatus;
 
 const FRAME_DELAY: Duration = Duration::from_millis(100);
